requirements.txt
psutil>=5.9
PyYAML>=6.0
rich>=13.0
Jinja2>=3.1
jsonschema>=4.19
